update all code from backend to frontend

+++ app/api/saran/route.js
@@ -1,9 +1,21 @@
 import { connectMongoDB } from "@/libs/mongodb";
-import GroupSaran from '@/models/groupSaranSchema';
-import Saran from "@/models/saranSchema";
-import User from "@/models/userSchema";
-import mongoose from "mongoose";
-import Vote from '@/models/voteSchema';
+import GroupSaran from '@/models/tes/groupSaranSchema';
+import Website from '@/models/tes/webSchema';
+import Saran from "@/models/tes/saranSchema";
+import Comment from "@/models/tes/commentSchema";
+import User from "@/models/tes/userSchema";
+import Vote from '@/models/tes/voteSchema';
+import { NextResponse } from 'next/server';
+import { validate as uuidValidate } from 'uuid';
+import nodemailer from 'nodemailer';
+
+const transporter = nodemailer.createTransport({
+    service: 'gmail',
+    auth: {
+        user: process.env.EMAIL_USER,
+        pass: process.env.EMAIL_PASSWORD
+    }
+});
 
 export async function GET(req) {
     await connectMongoDB();
@@ -11,18 +23,45 @@ export async function GET(req) {
     const link = url.searchParams.get('link');
     const id = url.searchParams.get('id');
     const sid = url.searchParams.get('sid');
+    const userId = url.searchParams.get('userId');
 
     try {
         let groupSaran;
 
         if (link) {
-            groupSaran = await GroupSaran.findOne({ link: link });
-        } else if (id && mongoose.Types.ObjectId.isValid(id)) {
-            groupSaran = await GroupSaran.findById(id);
-        } else if (sid && mongoose.Types.ObjectId.isValid(sid)) {
-            const saran = await Saran.findById(sid).populate('created_by', 'name profilePicture');
+            groupSaran = await Website.findOne({ link_advice: link }).lean();
+        } else if (id && uuidValidate(id)) {
+            groupSaran = await Website.findById(id);
+        } else if (userId && uuidValidate(userId)) {
+            const saranList = await Saran.find({ created_by: userId }).lean();
+
+            const saranListWithUsers = await Promise.all(
+                saranList.map(async (saran) => {
+                    const user = await User.findById(saran.created_by).select("name profilePicture");
+                    const group = await Website.findById(saran.groupSaranId).select("name");
+                    return {
+                        ...saran,
+                        created_by: user ? user.name : 'Unknown',
+                        group_name: group ? group.name : "Unknown Group"
+                    };
+                })
+            );
+
+            return new NextResponse(JSON.stringify({
+                success: true,
+                message: "Saran berhasil diambil berdasarkan ID pengguna",
+                data: saranListWithUsers,
+            }), {
+                status: 200,
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            });
+        }
+        else if (sid && uuidValidate(sid)) {
+            const saran = await Saran.findById(sid);
             if (!saran) {
-                return new Response(JSON.stringify({
+                return new NextResponse(JSON.stringify({
                     success: false,
                     message: "Saran tidak ditemukan",
                     data: null
@@ -32,13 +71,34 @@ export async function GET(req) {
             }
 
 
-            return new Response(JSON.stringify({ success: true, message: "Saran berhasil diambil", data: saran }), {
+            const user = await User.findById(saran.created_by);
+            if (!user) {
+                return new NextResponse(JSON.stringify({
+                    success: false,
+                    message: "User tidak ditemukan",
+                    data: null
+                }), {
+                    status: 404,
+                });
+            }
+
+            const saranWithUser = {
+                ...saran._doc,
+                created_by: user.name,
+                email: user.email,
+            };
+
+            return new NextResponse(JSON.stringify({
+                success: true,
+                message: "Saran berhasil diambil",
+                data: saranWithUser
+            }), {
                 status: 200,
             });
         } else {
-            return new Response(JSON.stringify({
+            return new NextResponse(JSON.stringify({
                 success: false,
-                message: "Parameter 'link' atau 'id' tidak valid",
+                message: "Parameter 'link', 'id', atau 'sid' tidak valid",
                 data: null
             }), {
                 status: 400,
@@ -46,7 +106,7 @@ export async function GET(req) {
         }
 
         if (!groupSaran) {
-            return new Response(JSON.stringify({
+            return new NextResponse(JSON.stringify({
                 success: false,
                 message: "Group Saran tidak ditemukan",
                 data: null
@@ -55,14 +115,35 @@ export async function GET(req) {
             });
         }
 
-        const saranList = await Saran.find({ groupSaranId: groupSaran._id }).populate('created_by', 'name profilePicture');
+        console.log('Group Saran:', groupSaran);
+        
+        const saranList = await Saran.find({ webId: groupSaran._id }).lean();
 
-        return new Response(JSON.stringify({ success: true, message: "Saran berhasil diambil", data: saranList }), {
+        const saranListWithUsers = await Promise.all(saranList.map(async (saran) => {
+            const user = await User.findById(saran.created_by);
+            const comments = await Comment.countDocuments({ saran_id: saran._id });
+            return {
+            ...saran,
+            created_by: user ? user.name : 'Unknown',
+            email: user ? user.email : 'Unknown',
+            profile_picture: user ? user.profilePicture : 'default_profile_picture_url',
+            comments: comments
+            };
+        }));
+        console.log('Saran List:', saranListWithUsers);
+
+        return new NextResponse(JSON.stringify({
+            success: true,
+            message: "Saran berhasil diambil",
+            data: {
+                ...saranListWithUsers,
+            }
+        }), {
             status: 200,
         });
     } catch (error) {
         console.error('Error:', error);
-        return new Response(JSON.stringify({ success: false, message: "Terjadi kesalahan saat mengambil data", error: error.message }), {
+        return new NextResponse(JSON.stringify({ success: false, message: "Terjadi kesalahan saat mengambil data", error: error.message }), {
             status: 500,
         });
     }
@@ -71,45 +152,49 @@ export async function GET(req) {
 export async function POST(req, res) {
     await connectMongoDB();
 
-    const url = new URL(req.url); // Ambil URL
+    const url = new URL(req.url);
     const link = url.searchParams.get('link');
-
-    const { title, description, userId } = await req.json();
+    const { title, description, userId, adminEmails } = await req.json();
 
     if (!link) {
-        return new Response(JSON.stringify({ message: "Parameter 'link' tidak ditemukan" }), {
+        return new NextResponse(JSON.stringify({ message: "Parameter 'link' tidak ditemukan" }), {
             status: 400,
         });
     }
 
     if (!userId) {
-        return new Response(JSON.stringify({ message: "User ID is required" }), {
+        return new NextResponse(JSON.stringify({ message: "User ID is required" }), {
             status: 400,
         });
     }
 
+    if (!adminEmails || !Array.isArray(adminEmails) || adminEmails.length === 0) {
+        return new NextResponse(JSON.stringify({ message: "Admin emails are required" }), { status: 400 });
+    }
+
     try {
-        // Cari group saran berdasarkan link
-        const groupSaran = await GroupSaran.findOne({ link: link });
+        const groupSaran = await Website.findOne({ link_advice: link });
+        console.log('Group Saran:', groupSaran);
+
         if (!groupSaran) {
-            return new Response(JSON.stringify({ message: "Group Saran tidak ditemukan" }), {
+            return new NextResponse(JSON.stringify({ message: "Group Saran tidak ditemukan" }), {
                 status: 404,
             });
         }
 
         const user = await User.findById(userId);
         if (!user) {
-            return new Response(JSON.stringify({ message: "User tidak ditemukan" }), {
+            return new NextResponse(JSON.stringify({ message: "User tidak ditemukan" }), {
                 status: 404,
             });
         }
 
-        // Ambil ID group saran
+
         const groupSaranId = groupSaran._id;
 
-        // Buat saran baru
+
         const newSaran = new Saran({
-            groupSaranId: groupSaranId,
+            webId: groupSaranId,
             title: title,
             description: description,
             saran: 'new',
@@ -119,7 +204,26 @@ export async function POST(req, res) {
 
         await newSaran.save();
 
-        return new Response(JSON.stringify({ success: true, data: newSaran }), {
+        const mailOptions = {
+            from: process.env.EMAIL_USER,
+            to: adminEmails.join(','),
+            subject: 'Notifikasi Saran Baru',
+            html: `
+                <div style="font-family: Arial, sans-serif; color: #333;">
+                    <h2 style="color: #4CAF50;">Notifikasi Saran Baru</h2>
+                    <p><strong>Judul:</strong> ${title}</p>
+                    <p><strong>Deskripsi:</strong> ${description}</p>
+                    <p><strong>Dibuat oleh:</strong> ${user.name}</p>
+                    <hr style="border: none; border-top: 1px solid #ddd;">
+                    <p style="font-size: 12px; color: #888;">Ini adalah email notifikasi otomatis. Mohon tidak membalas email ini.</p>
+                </div>
+            `
+        };
+
+
+        await transporter.sendMail(mailOptions);
+
+        return new NextResponse(JSON.stringify({ success: true, data: newSaran }), {
             status: 200,
         });
     } catch (error) {
@@ -128,17 +232,15 @@ export async function POST(req, res) {
     }
 }
 
-
 export async function PATCH(req) {
     await connectMongoDB();
     const { searchParams } = new URL(req.url);
     const id = searchParams.get("id");
-
-    // Ambil body dari request untuk mendapatkan status
     const body = await req.json();
-    const { status } = body; // Mengambil status dari body request
+    const { status, adminEmails } = body;
 
-    if (!mongoose.Types.ObjectId.isValid(id)) {
+
+    if (!uuidValidate(id)) {
         return new Response(
             JSON.stringify({
                 success: false,
@@ -152,8 +254,8 @@ export async function PATCH(req) {
     }
 
     try {
-        // Validasi status yang diizinkan
-        const allowedStatuses = ["new", "work in progress", "done", "cancelled"];
+
+        const allowedStatuses = ["new", "work in progress", "completed", "cancelled"];
         if (!allowedStatuses.includes(status)) {
             return new Response(
                 JSON.stringify({
@@ -166,12 +268,45 @@ export async function PATCH(req) {
             );
         }
 
-        // Cari dan perbarui status saran berdasarkan ID
+        // if (!adminEmails || !Array.isArray(adminEmails) || adminEmails.length === 0) {
+        //     return new NextResponse(JSON.stringify({ message: "Admin emails are required" }), { status: 400 });
+        // }
+
         const updatedSaran = await Saran.findByIdAndUpdate(
             id,
-            { status: status, updated_at: Date.now() }, // Mengubah status dan waktu update
-            { new: true } // Mengembalikan data yang baru setelah update
+            { status: status, updated_at: Date.now() },
+            { new: true }
         );
+
+        const saran = await Saran.findById(id);
+        if (!saran) {
+            return new Response(
+                JSON.stringify({
+                    success: false,
+                    message: "Saran not found",
+                }),
+                {
+                    status: 404,
+                }
+            );
+        }
+
+        const mailOptions = {
+            from: process.env.EMAIL_USER,
+            to: adminEmails,
+            subject: 'Notifikasi Pembaruan Status Saran',
+            html: `
+            <div style="font-family: Arial, sans-serif; color: #333;">
+                <h2 style="color: #4CAF50;">Notifikasi Pembaruan Status Saran</h2>
+                <p><strong>Judul Saran:</strong> ${saran.title}</p>
+                <p><strong>Status Baru:</strong> ${status}</p>
+                <hr style="border: none; border-top: 1px solid #ddd;">
+                <p style="font-size: 12px; color: #888;">Ini adalah email notifikasi otomatis. Mohon tidak membalas email ini.</p>
+            </div>
+            `
+        };
+
+        await transporter.sendMail(mailOptions);
 
         if (!updatedSaran) {
             return new Response(
@@ -208,7 +343,7 @@ export async function DELETE(req) {
     const { searchParams } = new URL(req.url);
     const id = searchParams.get("id");
 
-    if (!mongoose.Types.ObjectId.isValid(id)) {
+    if (!uuidValidate(id)) {
         return new Response(JSON.stringify({
             success: false,
             message: "Invalid ID format",
@@ -234,13 +369,18 @@ export async function DELETE(req) {
             });
         }
 
+        const votes = await Vote.find({ saranId: id });
+
+        if (votes.length > 0) {
+            await Vote.deleteMany({ saranId: id });
+        }
+
         await Saran.findByIdAndDelete(id);
-        await Vote.deleteMany({ saranId: id });
 
         return new Response(JSON.stringify({ message: "Advice and related votes deleted" }), {
             status: 200,
         });
-        
+
     } catch (error) {
         console.error(error.message);
         return new Response(JSON.stringify({
